requirements.txt
autopep8==1.6.0
cffi==1.15.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
toml==0.10.2
